forbidden funciona

+++ Angular/src/app/_services/user-auth.service.ts
@@ -12,15 +12,15 @@ export class UserAuthService {
             this.localStorage = this.document.defaultView?.localStorage;
          }
 
-    private setTime() {
-        if (localStorage) {
+    public setTime() {
+        if (this.localStorage) {
             const expirationTime = Date.now() + 2 * 60 * 60 * 1000;
             this.localStorage.setItem("expiration_time", expirationTime);
         }
     }
 
-    private getTime() {
-        if (localStorage) {
+    public getTime() {
+        if (this.localStorage) {
             const expirationTime = this.localStorage.getItem("expiration_time");
             // const expirationTime = sessionStorage.getItem("expiration_time");
             if (Date.now() <= expirationTime) {
@@ -29,25 +29,21 @@ export class UserAuthService {
                 return false;
             }
         }
-        return;
+        return false;
     }
 
-    public setRole(roles: string) {
+    public setRole(role: string) {
         if (localStorage) {
-            this.localStorage.setItem("role", JSON.stringify(roles));
-            // sessionStorage.setItem("role", JSON.stringify(roles));
+            this.localStorage.setItem("role", JSON.stringify(role));
+            // sessionStorage.setItem("role", JSON.stringify(role));
         }
     }
 
     public getRole(): string {
         if (localStorage) {
-            if(this.getTime()) {
                 const storedRole = this.localStorage.getItem("role");
                 // const storedRole = sessionStorage.getItem("role");
                 return storedRole ? JSON.parse(storedRole) : '';
-            } else {
-                this.clear();
-            }
         }
         return '';
     }
@@ -61,12 +57,8 @@ export class UserAuthService {
 
     public getToken(): string {
         if (localStorage) {
-            if(this.getTime()) {
                 return this.localStorage.getItem("jwtToken") ?? '';
                 // return sessionStorage.getItem("jwtToken") ?? '';
-            } else {
-                this.clear();
-            }
         }
         return '';
     }
@@ -80,25 +72,26 @@ export class UserAuthService {
 
     public getNome(): string {
         if (localStorage) {
-            if(this.getTime()) {
                 return this.localStorage.getItem("nome") ?? '';
                 // return sessionStorage.getItem("nome") ?? '';
-            } else {
-                this.clear();
-            }
         }
         return '';
     }
 
     public clear() {
-        this.localStorage.clear();
-        // sessionStorage.clear();
-        this.router.navigate(['/home']).then(() => {
-          window.location.reload();
-        });
+        if (this.localStorage) {
+            this.localStorage.clear();
+            // sessionStorage.clear();
+            this.router.navigate(['/home']).then(() => {
+            window.location.reload();
+            });
+        }
     }
 
     public isLoggedIn() {
-        return this.getRole() && this.getToken();
+        if (this.localStorage && this.getTime()) {
+            return this.getRole() && this.getToken();
+        }
+        return false;
     }
 }
